% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{collect_panel_css}
\alias{collect_panel_css}
\title{Collect CSS strings from a set of panels}
\usage{
collect_panel_css(panel_names)
}
\arguments{
\item{panel_names}{Character vector of panel names}
}
\value{
Concatenated CSS string (empty string if no panels have CSS)
}
\description{
Collect CSS strings from a set of panels
}
\keyword{internal}
