% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{collect_panel_js}
\alias{collect_panel_js}
\title{Collect JavaScript strings from a set of panels}
\usage{
collect_panel_js(panel_names)
}
\arguments{
\item{panel_names}{Character vector of panel names}
}
\value{
Concatenated JS string (empty string if no panels have JS)
}
\description{
Collect JavaScript strings from a set of panels
}
\keyword{internal}
