% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_module_feature.R
\name{.build_feature_report_module}
\alias{.build_feature_report_module}
\title{Build the Feature Diagnostics report module}
\usage{
.build_feature_report_module(available, active)
}
\arguments{
\item{available}{Whether Feature Diagnostics data are available.}

\item{active}{Whether this is the initially active module.}
}
\value{
A report module specification.
}
\description{
Build the Feature Diagnostics report module
}
\keyword{internal}
