% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_module_pca.R
\name{.build_pca_report_module}
\alias{.build_pca_report_module}
\title{Build the PCA report module}
\usage{
.build_pca_report_module(available, has_sample)
}
\arguments{
\item{available}{Whether PCA data are available.}

\item{has_sample}{Whether sample colouring is available.}
}
\value{
A report module specification.
}
\description{
Build the PCA report module
}
\keyword{internal}
