% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_module_qc.R
\name{.build_qc_report_module}
\alias{.build_qc_report_module}
\title{Build the QC report module}
\usage{
.build_qc_report_module(available)
}
\arguments{
\item{available}{Whether QC data are available.}
}
\value{
A report module specification.
}
\description{
Build the QC report module
}
\keyword{internal}
