% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data_ports.R
\name{.build_report_data_ports}
\alias{.build_report_data_ports}
\title{Build embedded data and script tags for registered report modules}
\usage{
.build_report_data_ports(
  marker_json,
  clusters_json,
  cluster_counts_json,
  marker_n_top,
  dim_opacity,
  has_samples,
  sample_comp_json,
  cluster_colors_json,
  gene_expr_json,
  marker_gene_clusters_json,
  all_gene_sources_json,
  has_plot,
  qc_payload,
  has_feature,
  feature_diag_json,
  use_webgl,
  has_pca,
  pca_data_json,
  pca_has_sample,
  pca_color_by,
  pca_all_pcs_json,
  pca_loading_json,
  pca_loading_top_n,
  first_view,
  has_umap,
  panel_js_extra
)
}
\value{
A list of script tags in dependency order.
}
\description{
Build embedded data and script tags for registered report modules
}
\keyword{internal}
