% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_module_contract.R
\name{.first_report_module}
\alias{.first_report_module}
\title{Find the first available report module}
\usage{
.first_report_module(modules)
}
\arguments{
\item{modules}{List of report module specifications.}
}
\value{
Module id.
}
\description{
Find the first available report module
}
\keyword{internal}
