% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_module_contract.R
\name{.render_report_module_tabs}
\alias{.render_report_module_tabs}
\title{Render module tabs}
\usage{
.render_report_module_tabs(modules, first_view)
}
\arguments{
\item{modules}{List of report module specifications.}

\item{first_view}{Id of the initially active module.}
}
\value{
View-tab tag.
}
\description{
Render module tabs
}
\keyword{internal}
