% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_assets.R
\name{.report_asset_path}
\alias{.report_asset_path}
\title{Locate a packaged report asset}
\usage{
.report_asset_path(...)
}
\arguments{
\item{...}{Path components below inst/assets.}
}
\value{
Absolute path to the requested asset.
}
\description{
Locate a packaged report asset
}
\keyword{internal}
