% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{html_escape}
\alias{html_escape}
\title{Escape text before inserting it into report HTML}
\usage{
html_escape(x)
}
\arguments{
\item{x}{Values used in an HTML text node.}
}
\value{
Escaped character vector.
}
\description{
Escape text before inserting it into report HTML
}
\keyword{internal}
