% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scReportLite-package.R
\docType{package}
\name{scReportLite-package}
\alias{scReportLite}
\alias{scReportLite-package}
\title{scReportLite package}
\description{
Interactive report generation for pre-computed single-cell results.
}
\author{
\strong{Maintainer}: Kee-gong Karios Park \email{karioslion0517@gmail.com}

}
\keyword{internal}
