% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_sc_report_parameters}
\alias{validate_sc_report_parameters}
\title{Validate public sc_report scalar parameters}
\usage{
validate_sc_report_parameters(
  cluster_col,
  cell_col,
  sample_col,
  pca_color_by,
  pca_loading_top_n,
  output,
  title,
  point_size,
  point_alpha,
  dim_opacity,
  marker_n_top,
  panels,
  use_webgl
)
}
\description{
Validate public sc_report scalar parameters
}
\keyword{internal}
