% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_assembly.R
\name{assemble_report}
\alias{assemble_report}
\title{Assemble and write the complete HTML report}
\usage{
assemble_report(
  umap_plot = NULL,
  umap_df = NULL,
  marker_df,
  cluster_col,
  cell_col,
  sample_col,
  gene_expr_df = NULL,
  pca_df = NULL,
  pca_data_json = "null",
  pca_has_sample = FALSE,
  pca_color_by = "cluster",
  pca_all_pcs_json = "[]",
  pca_loading_json = "[]",
  pca_loading_top_n = 10,
  qc_payload = NULL,
  feature_diag = NULL,
  feature_diag_json = "null",
  use_webgl = TRUE,
  output,
  title,
  dim_opacity,
  marker_n_top,
  panels = c("umap", "marker_table")
)
}
\arguments{
\item{umap_plot}{A plotly htmlwidget object}

\item{umap_df}{The input UMAP data frame (for sidebar stats)}

\item{marker_df}{The input marker data frame (NULL or data.frame)}

\item{cluster_col}{Name of the cluster column}

\item{cell_col}{Name of the cell column}

\item{sample_col}{Optional name of the sample column in umap_df (NULL to skip)}

\item{output}{Path to output HTML file}

\item{title}{Report title}

\item{dim_opacity}{Opacity for non-highlighted points (0-1)}

\item{marker_n_top}{Number of top marker genes to show per cluster}
}
\value{
Invisibly, the path to the output file
}
\description{
Assemble and write the complete HTML report
}
\keyword{internal}
